+++ NEWS.md
@@ -1,5 +1,7 @@
 # ritest 0.0.1 (in development)
 
+* Add tidiers for compatability with `broom`.
+
 * Add `h0` argument for specifying the null hypothesis (replaces the previous 
 `pvals` argument and takes on different functionality)